programmer and projects

+++ src/main/java/com/bharath/springdata/associations/manytomany/repository/ProgrammerRepository.java
@@ -3,5 +3,5 @@ package com.bharath.springdata.associations.manytomany.repository;
 import com.bharath.springdata.associations.manytomany.entities.Programmer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProgrammerRepository extends JpaRepository<Programmer, Long> {
+public interface ProgrammerRepository extends JpaRepository<Programmer, Integer> {
 }
